requirement.txt
flask==1.1.2 
numpy==1.19.5
flask-cors==3.0.10
pandas==1.1.5
flask-restful==0.3.8
scikit-learn-0.24.1
sklearn-0.0
pymongo==3.11.0